backend-common: move HostDiscovery to backend-app-api
