+++ .changeset/calm-carpets-kiss.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-common': patch
+---
+
+The `HostDiscovery` export has been deprecated, import it from `@backstage/backend-app-api` instead.
