+++ .changeset/serious-penguins-tease.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-test-utils': patch
+---
+
+Updated to import `HostDiscovery` from `@backstage/backend-app-api`.
