+++ .changeset/six-weeks-raise.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-app-api': patch
+---
+
+Moved `HostDiscovery` from `@backstage/backend-common`.
