+++ packages/backend-app-api/src/services/implementations/discovery/discoveryServiceFactory.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import { HostDiscovery } from '@backstage/backend-common';
 import {
   coreServices,
   createServiceFactory,
 } from '@backstage/backend-plugin-api';
+import { HostDiscovery } from './HostDiscovery';
 
 /** @public */
 export const discoveryServiceFactory = createServiceFactory({
